Create 36 - Valid Sudoku
Determine if a Sudoku is valid, according to Sudoku Puzzles Rules.
The Sudoku board could be partially filled, where empty cells are filled with the character '.'.
A partially filled sudoku which is valid.

Note:
A valid Sudoku board (partially filled) is not necessarily solvable. Only the filled cells need to be validated.